Add user info to tasks, and make them available only for owners

+++ app/resources/task/delete-task.ts
@@ -2,7 +2,7 @@ import { BadRequest } from '@curveball/http-errors';
 import { taskTable } from '@infrastructure/dynamodb';
 
 import { Event, Context } from '@app/helpers';
-import { DynamoDB } from '@app/database';
+import DynamoDB from '@app/helpers/dynamodb';
 
 export default {
   path: 'DELETE /task/{id}',
@@ -10,12 +10,12 @@ export default {
   lambda,
 };
 
-async function lambda(ev: Event, ctx: Context) {
+async function lambda(ev: Event, ctx: Context, userId: string) {
   if (!ev.pathParameters?.id) {
     throw new BadRequest('Missing id');
   }
 
   const { id } = ev.pathParameters;
 
-  return DynamoDB.delete(taskTable.name.get(), id);
+  return DynamoDB.delete(taskTable.name.get(), { id, userId });
 }
